Add unit directory; move mock tests
