+++ test/unit/util_test.go
@@ -1,0 +1,12 @@
+package unit
+
+import (
+	"testing"
+
+	"github.com/linode/linodego"
+	"github.com/linode/linodego/internal/testutil"
+)
+
+func createMockClient(t *testing.T) *linodego.Client {
+	return testutil.CreateMockClient(t, linodego.NewClient)
+}
